Redo project structure with separate library in src
Fixes VSCode errors and allows linting to work. Also comment out gloss dependency cause it breaks repl.
