--- app/model/General.hs
@@ -1,4 +1,0 @@
-module Model.General where
-  type Position = (Int, Int)
-  type Vector   = (Int, Int)
-  type Score    = Int
