--- app/model/PowerUp.hs
@@ -1,7 +1,0 @@
-module Model.PowerUp where
-
-  import Model.General (Position)
-    
-  data PowerUp = BurstFire {burstFirePos :: Position}
-               | ConeFire {coneFirePos :: Position}
-               | SpeedBoost {speedBoostPos :: Position}
