+++ shootemup.cabal
@@ -1,23 +1,39 @@
 cabal-version:      2.4
 name:               shootemup
 version:            0.1.0.0
+
 author:             Bram Boogaarts, Yanna van Straten
 maintainer:         bram.boogaarts@gmail.com
 
-executable Game
-  main-is:          Main.hs
-
-  other-modules:    Model.Enemy,
+library
+  exposed-modules:  Model.Enemy,
                     Model.GameState,
                     Model.General,
                     Model.Player,
-                    Model.Shooting
+                    Model.Shooting,
                     Model.PowerUp
+
+  -- Modules included in this library but not exported.
+  -- other-modules:
 
   -- LANGUAGE extensions used by modules in this package.
   -- other-extensions:
-  build-depends:    base ^>=4.17.2.0,
-                    gloss,
-                    random
+  build-depends:    base ^>=4.17.2.0
+                  , random
+                  -- , gloss
+  hs-source-dirs:   src
+  default-language: Haskell2010
+
+executable shootemup
+  main-is:          Main.hs
+
+  -- Modules included in this executable, other than Main.
+  -- other-modules:
+
+  -- LANGUAGE extensions used by modules in this package.
+  -- other-extensions:
+  build-depends:    base ^>=4.17.2.0
+                  , shootemup
+
   hs-source-dirs:   app
   default-language: Haskell2010
