+++ src/model/General.hs
@@ -1,0 +1,5 @@
+module Model.General where
+
+type Position = (Int, Int)
+type Vector   = (Int, Int)
+type Score    = Int
