+++ src/model/PowerUp.hs
@@ -1,0 +1,8 @@
+module Model.PowerUp where
+
+import Model.General (Position)
+  
+data PowerUp = BurstFire {burstFirePos :: Position}
+              | ConeFire {coneFirePos :: Position}
+              | SpeedBoost {speedBoostPos :: Position}
+
